refactor: 기존 RoomService 클래스에서 모임방 참여 관련 기능을 분리
